requirements.txt
numpy==1.24.4
pandas==2.0.3
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
sqlalchemy==2.0.23
